refactor: testを適切な場所に移動
