+++ packages/rust-core/crates/engine-core/src/search/mod.rs
@@ -12,8 +12,6 @@ pub mod types;
 pub mod unified;
 
 #[cfg(test)]
-pub mod prefetch_test;
-#[cfg(test)]
 pub mod test_helpers;
 
 // Re-export commonly used items
